Add files via upload
This is my .hmk memory channel file from the Kenwood TH-D74A/E Memory Control Program MCP-D74. It is basically a csv file that the script uses to match a memory channel number (which it receives over the USB serial connection) to the memory channel name (which is not available over the USB serial connection). You are welcome to use it or just review it for background, but unless you have your memory frequencies programmed the same way with the same names, it won't match your programming.